Changed global dataloader to dataloader per request
